Progress towards user story https://git.cardiff.ac.uk/c21052928/energy-data-centre-visualisation/-/issues/6 I have moved all the complex calculation logic that was being done with the data on the front-end to the API.
Data will be read from redis. I only read the needed columns for the calculations from the redis cache into memory.

I am using polars and it's lazy loading functionality to efficiently query the data. Which is then used to create a json object to send to the front-end that contains all the needed information for d3.js to generate a clustered bar chart.

This implementation presumes that the initalisation python file has been run prior.

Code is commented throughout.