Add API route to load map data
We currently have a hardcoded set of map names, with relative csv paths to the dataset folder.

We have a custom processed and simplified shapefile that is loaded with geopandas, and is joined with the csv dataframe. This is really slow and hence takes ~6s for a request to complete.

In the future, we need to find a way to do everything with polars, and could further simplify the processing to make it even faster.