Myself and Jack have worked on transferring the logic that we previousy created to aggregate the data used in the pie chart, to the backend.
Progress towards user story https://git.cardiff.ac.uk/c21052928/energy-data-centre-visualisation/-/issues/7

We pull the data from redis, specifying the exact columns, to save on memory. Polars is used to sort, and manipulate the data/logic.

Ultimately a list of heating types alongside their mean averages are produced and returned to the front-end to be rendered in the form of a pie chart.

Co-authored-by: Jack Richards <richardsj31@cardiff.ac.uk>
Signed-off-by: c21087406 <alsarmih@cardiff.ac.uk>